Add changes requested by the reviewer
